fix(telegram): keep audio preflight checks typed
