+++ extensions/telegram/src/bot-message-context.typing.test.ts
@@ -83,7 +83,7 @@ describe("buildTelegramMessageContext typing", () => {
         text: undefined,
         voice: { file_id: "voice-1", duration: 1 },
       },
-      allMedia: [{ path: "/tmp/voice.ogg", contentType: "audio/ogg" }],
+      allMedia: [{ kind: "audio", path: "/tmp/voice.ogg", contentType: "audio/ogg" }],
       sendChatActionHandler,
     });
 
@@ -121,7 +121,7 @@ describe("buildTelegramMessageContext typing", () => {
         text: undefined,
         voice: { file_id: "voice-1", duration: 1 },
       },
-      allMedia: [{ path: "/tmp/voice.ogg", contentType: "audio/ogg" }],
+      allMedia: [{ kind: "audio", path: "/tmp/voice.ogg", contentType: "audio/ogg" }],
       sendChatActionHandler,
     });
